pyproject.toml
[tool.poetry]
name = "aact_openhands"
version = "0.0.5"
description = "Openhands integrated with aact"
authors = ["Arpandeep Khatua <arpandeepk@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = []
packages = [
    { include = "aact_openhands" },
    { include = "openhands", from = "openhands" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
aact = "*"
rich = ">=13.6.0,<14.0.0"
python-dotenv = "^1.0.0"
flask = "^3.0.0"
flask-cors = "^4.0.0"
pydantic = "^2.6.1"
websockets = "^12.0.0"
importlib-resources = "^6.1.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
ruff = "^0.0.262"
pytest = "^7.4.3"
requests = "^2.31.0"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core >= 1.0.0"]
build-backend = "poetry.core.masonry.api"